encapsulate CountrySelector implementation by introducing an interface
